# Conflicts:
#	README.md
#	src/main/scala/com/raquo/airstream/web/AjaxEventStream.scala
#	src/main/scala/com/raquo/airstream/web/DomEventStream.scala

+++ src/test/scala/com/raquo/airstream/AsyncUnitSpec.scala
@@ -2,7 +2,6 @@ package com.raquo.airstream
 
 import org.scalatest.funspec.AsyncFunSpec
 import org.scalatest.matchers.should.Matchers
-
 import scala.concurrent.{ExecutionContext, Future, Promise}
 import scala.scalajs.js
 import scala.util.Try
@@ -17,7 +16,7 @@ class AsyncUnitSpec extends AsyncFunSpec with Matchers {
 
   def delay[V](millis: Int)(value: => V): Future[V] = {
     val promise = Promise[V]()
-    js.timers.setTimeout(millis) {
+    js.timers.setTimeout(millis.toDouble) {
       promise.complete(Try(value))
     }
     promise.future.map(identity)(executionContext)
